+++ CHANGELOG.md
@@ -15,6 +15,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+
+## [2.0.4]
+
+### Changed
+
+- Remove Sync marker in Tx, Rx, AsyncTx, AsyncRx to prevent misuse with Arc
+
+
 ## [2.0.3] - 2025-07-07
 
 ### Changed
@@ -25,7 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - AsyncRx should not have Clone.
 
-- Protect against mis-use of spsc/mpsc when user should use mpmc (avoiding deadlocks)
+- Protect against misuse of spsc/mpsc when user should use mpmc (avoiding deadlocks)
 
 ## [2.0.2] - 2025-07-05